finish

+++ products/delete.php
@@ -5,16 +5,21 @@
     $image_name = filterGetRequest("image");
 
     $stmt = $con->prepare("DELETE FROM products WHERE id = ? and user_id=?");
-    $stmt->execute(array($produ_id, $user_id));
+    $stmt->execute(array($product_id, $user_id));
     
     $count = $stmt->rowCount();
     
+    
     if ($count > 0) {
-        if(file_exists("../assets/images/".$image_name)) {
-            deleteFile("../assets/images", $image_name);
+        if($image_name){
+            if(file_exists("../assets/images/".$image_name)) {
+                deleteFile("../assets/images", $image_name);
+            }
         }
         echo json_encode(array("status" => "success", "message" => "deleted successfully"));
-    } else {
-        echo json_encode(array("status" => "fail", "message" => "fail !!"));
-    }    
+        return;
+    }
+    
+    echo json_encode(array("status" => "fail", "message" => "fail !!"));
+        
 ?>
